adding convolution layers
the goal of this is to increase the receptive field of our neurones, this can be limiting if we do not have enough data to harness the accumlation of information added with the gain of the receptive field. Luckily having stong data augmentation we can double and even add 4 orders of magnitude to our data to harness full potential of what the data gives us.